Got my color computation right*
Basically, for any given *temperature*, we figure out how far away it
is from ideal and consider that the "base" color.  Then, for any
given *pixel*, we figure out what its relevance is based on distance
from the sensor and figure out what color it is between pure base and
pure white.